feat: moved listeners to components
