+++ src/main.js
@@ -1,7 +1,8 @@
 import ListFilterView from './view/list-filter-view.js';
 import TripPresenter from './presenter/trips-presenter.js';
-import {render} from './render.js';
+//import {render} from './render.js';
 import PointsModel from './model/points-model.js';
+import {render} from './framework/render.js';
 
 const headerFiltersElement = document.querySelector('.trip-controls__filters');
 const mainEventsElement = document.querySelector('.trip-events');
