+++ src/view/edit-point-view.js
@@ -121,14 +121,33 @@ const createEditPointTemplate = (point) => {
 
 export default class EditPointView extends AbstractView {
   #point = null;
+  #handleFormSubmit = null;
+  #handleRollupBtnClick = null;
 
-  constructor({point}) {
+  constructor({point, onFormSubmit, onRollupBtnClick}) {
     super();
     this.#point = point;
+    this.#handleFormSubmit = onFormSubmit;
+    this.#handleRollupBtnClick = onRollupBtnClick;
+
+    this.element.querySelector('form')
+      .addEventListener('submit', this.#formSubmitHandler);
+    this.element.querySelector('.event__rollup-btn')
+      .addEventListener('click', this.#rollupButtonClickHandler);
+
   }
 
   get template() {
     return createEditPointTemplate(this.#point);
   }
 
+  #formSubmitHandler = (evt) => {
+    evt.preventDefault();
+    this.#handleFormSubmit();
+  };
+
+  #rollupButtonClickHandler = (evt) => {
+    evt.preventDefault();
+    this.#handleRollupBtnClick();
+  };
 }
