+++ src/view/point-view.js
@@ -61,14 +61,23 @@ const createPointTemplate = (point) =>{
 
 export default class PointView extends AbstractView {
   #point = null;
+  #handleRollupBtnClick = null;
 
-  constructor({point}) {
+  constructor({point, onRollupBtnClick}) {
     super();
     this.#point = point;
+    this.#handleRollupBtnClick = onRollupBtnClick;
+
+    this.element.querySelector('.event__rollup-btn')
+      .addEventListener('click', this.#rollupBtnClickHandler);
   }
 
   get template() {
     return createPointTemplate(this.#point);
   }
 
+  #rollupBtnClickHandler = (evt) => {
+    evt.preventDefault();
+    this.#handleRollupBtnClick();
+  };
 }
